Adjust defining t_size as size_t

+++ libft.h
@@ -6,14 +6,14 @@
 /*   By: robernar <robernar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/16 23:08:08 by robernar          #+#    #+#             */
-/*   Updated: 2023/11/14 08:03:38 by robernar         ###   ########.fr       */
+/*   Updated: 2023/11/14 08:44:19 by robernar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef LIBFT_H
 # define LIBFT_H
 # include <stdlib.h>
 
-typedef long	t_size;
+typedef size_t	t_size;
 int		ft_isalpha(int c);
 int		ft_isdigit(int c);
 int		ft_isalnum(int c);
